+++ Sources/KeyboardKit/_Deprecated/Services+Deprecated.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "Keyboard.StandardBehavior")
+public typealias StandardKeyboardBehavior = Keyboard.StandardBehavior